V2.1- Looping de execução da IA + criação Regras 1 da IA.    agora ao iniciar o endpoint de "ligar coleira" ele ficar utilizando a IA em loop ate que a coleira seje desligada. Criação das regras de nivel 1 da IA, basicamente verificam se é pra desligar a coleira ou nao.

+++ src/main/java/com/example/demo/repo/AnimalRepo.java
@@ -7,5 +7,5 @@ import org.springframework.data.rest.webmvc.RepositoryRestController;
 @RepositoryRestController
 public interface AnimalRepo extends JpaRepository<AnimalModel, Long> {
 
-    AnimalModel findAnimalModelBycoleiraid(int coleiraid);
+    AnimalModel findAnimalModelBycoleiraid(long coleiraid);
 }
